requirements.txt
django
django-recaptcha
django-maintenancemode
django-redis
channels
asgi_redis
daphne
python-dateutil
dill
beautifulsoup4
lxml
requests